openhpi-visualizer/requirements.txt
requests
beautifulsoup4
langfun
google-generativeai
python-dotenv
pyglove
plotly
